requirements.txt
google-generativeai 
python-pptx
python-dotenv
streamlit
python-docx
requests
beautifulsoup4
pyrebase4


